requirements-windows.txt
pyserial
Flask
numpy
# Needed if you are building binary distributions
PyInstaller

